requirements.txt
django
requests
line-bot-sdk
redis
celery
uwsgi
python-dotenv